+++ src/app.cc
@@ -177,7 +177,7 @@ int App::handle_custom_event(int param_one, int param_two) {
     const std::string location{raw_location->data()};
     this->search_location(location);
     return 1;
-  } else if (param_one == CustomEvent::display_alert) {
+  } else if (param_one == CustomEvent::open_alert_viewer) {
     if (this->ui) {
       this->ui->display_alert();
       return 1;